requirements.txt
Flask==2.3.3
psycopg-binary==3.2.10
python-dotenv==1.0.0
Pillow==11.3.0
requests==2.32.5
numpy==2.3.3
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.3.0
itsdangerous==2.2.0
typing_extensions==4.15.0
setuptools==80.9.0
wheel==0.45.1
